experiments using graphviz.
The graph here is a fairly practical example of things I'm actually doing
in my desktop working environment right now: capturing some stuff from
other upstream package managers, using Timeless Wares to keep an immutable
version of it, and then running.

This is just poking around with how efficiently we can hint to graphviz
some of the features of our problem space.  Turns out, a fair bit:
using subgraphs and (critically; I never know about these before!)
`constraint=false` attributes, we can actually get a layout that does
a fairly reasonable job of showing independent sections.

It still starts clunking pretty hard if you stretch it to more data though.
This example is one of the accidentally-nice ones.  I've gotten some
*very* ugly stuff to come out with arrows careening through nodes, particularly
on the constraint=false lines that head back to the left.

One thing I've utterly failed at so far is figuring out a way have a bunch of
discrete output cites on a formula node... can graphviz do that?  Probably,
you'd think, but I'm at a loss to find the incantation so far (hacking it in
with subgraphs is *real* messy and basically doesn't fly).

We might continue to run with stuff like this for a while, and asking
for only a limited number of subgraphs to be rendered at one time might produce
pretty valid results.  It's also probably valid to set some limits on our
maximum acceptable sunk costs with graphviz workarounds though -- building our
own toposort and layout is an investment that's well within reach.  We *do*
have domain specific knowledge, and the ability to choose to break long arcs,
etc, and that makes a big impact on how good of visualization we can produce.

Signed-off-by: Eric Myhre <hash@exultant.us>