Draft an entirely more radical use of skylark.
At some point we have to get away from continuing to sling json around.
Maybe something like this is the way forward?

It's hard to figure out where exactly we should get serious about types.
The amount of coding overhead for implementing core things in the golang side
is still pretty high unless we put some serious thought into boilerplate
reduction on that end.  And arguably, it should be a *feature* if we can
dogfood even our most commonly used composition operators in the skylark?

Signed-off-by: Eric Myhre <hash@exultant.us>