Add reified ReleaseItemID to skylark.
I tried drafting things with plain strings in the skylark side alone, and while
that's possible, it didn't seem pretty, and would leave any user code that
wants to work with the structure as a tuple stuck re-implementing some messy
stringy stuff.  A custom struct here is worth it.

(I wonder to what extent reviewing the skylarkstruct package from upstream
would be useful for this.  Possibly some.  But I'm not afraid of investing in
our own manually reified types either; it's just an opportunity to build
particularly good error and usage messaging.)

Signed-off-by: Eric Myhre <hash@exultant.us>