placeholder ReleaseItemID resolvers.
These will need *some* additional sprinkling of fairy dust to be fully
implemented, because they'll need to actually *read data* from a hitch DB,
and that requires some departure from the Skylark sandbox.

What API that will have is TBD.  Maybe we should implement those things in
blessed functions, just because there's no sense in having a huge number
of fragmented approaches to version resolution in the core of the community.
Or maybe we'll put in some basic reader functions that will readly dump
you the entire hitch DB content and let you decide.  Design: later.

(It's been long enough since this whole skylark file compiled and ran
that I really want to get back on the happy path with it before dealing
with any of these hard things on the edges.)

Signed-off-by: Eric Myhre <hash@exultant.us>