Use go-wish to tighten up some tests.
The line count decreases only slightly, but we're just getting started.

The fixture code is now more readable, because we can use indentation
in the source code freely and still get the right strings.

The output in case of failures will also now be better: we'll get
diffs, and such!

The 'mustGlobalKeys' function remains a bit interesting and custom,
because it needs to unpack the unique skylark StringDict type.
But, we'll make further improvements to this in the next commit.

Signed-off-by: Eric Myhre <hash@exultant.us>