First pass on logistics for even bigger graphs.
Basting, it turns out, is not enough.  It's useful to string small
locally-grouped sets of formulas together for making moderately complex
products (and making it possible to run their tests in separately
isolated environments, lets not forget).  But a bigger picture remains.

"Commissions" (working title, like much of this -- the nomenclature
refactor date is probably not far off now) are meant to be this bigger
picture.

Commissions will be sort of like an ultra-basting: by looking at a
group of bastings and doing a topological sort on their imports, we
can build a DAG of which things would be evaluted if building the
entire suite of bastings at once from their farthest-back sources.

We're going to give this first class support: it will be possible to
calculate the commission graph for an entire set of projects, and
importantly, to do it *even without building anything*.  We're going
to do this without significantly new data types; we can actually get
everything to work within our existing types, and in fact without
changing much of the model of "planners" being any ol' abstract piece
of logic that can emit a basting.  By blackboxing the planner and
controlling the presentation of release info it sees, we can actually
check if the planner would accept hypothetical releases without
knowing anything about the logic it uses internally to decide that.
This will allow us to compute downstream splash radiuses of any changes
anywhere in the commission graph; provide the whole commission graph
and a toposort order of it to an evaluation agent; and all sorts of
other excellent things like that.

This also marks a change in thinking about where exactly the halting
points in the system are.  Of course, release records will still be
serializable points, and the replay instructions in them expected to
be, well, totally capable of supporting replay of that build process.
However, rather than visualizing our whole system as a bunch of basting
processes connecting release cells, it's now more advisable to
visualize the whole commission graph as a bunch of conjoined bastings,
and the release cells are actually precipitants off of that graph.
They're deeply related; but the commission graph exists before and
independent of the release cells.  And it's purely "coincidental" that
re-examining a huge heap of release records and their attached replay
instructions later in tinfoil-hat audit mode will *happen* to draw up
the same graph as the commission originally set out to build up.

There's a LOT of "small matter of programming" between the present
versus realizing that whole picture, of course.  This commit is just
a fraction of the work ahead.

Signed-off-by: Eric Myhre <hash@exultant.us>