commission: graph explore has plural node kinds.
So, here's a union to deal with that, in brief.

Vacilating on many details, discussion in comments.

There are certainly several off-by-ones in recurison; ignoring until
other bits get full-formed enough to wrap whole thing in test cases.

Started test and test mock outline.

Biggest question at the moment is probably... how worth it *is* it to
try to get this commission graph exploration to be free of details of
our skylark-based scripting concepts?  We have the theorycraft to make
the planner an incidental, but also, do I want to add modes to this
graph node kind union which will allow e.g. exec'ing another planner
freely?  I'm substantially dubious about that; many many people have
given the feedback that their interest in the system as a whole goes
down sharply if there's an unbounded turing-complete thing running
during planning that they now have to worry about.  So... we can do it,
but is it worth it?  Probably not, for this prototype phase, at least.

Some comments may be half-baked and missing their closing parens...
this commit is firmly in "showing the making of the sausage" territory.
(Also 3am coding is bad sometimes mkay?)

Signed-off-by: Eric Myhre <hash@exultant.us>