layout,commission: new pkg to load module cfg.
A fairly standard file layout is going to be a necessary convention.

And it's not just commissioning large builds that's going to look for
this file layout; we're also going to need an ecosystem of tools that
perform functions like looking for any catalog snippets in your tree so
that we can then look at public resources and see if they're up to
date, etc -- that'll need to inspect for all the same sorts of stuff.
(Might not need to load the build script into memory, but we'll cross
bridges like that one when we get to it.)

Signed-off-by: Eric Myhre <hash@exultant.us>