wire up a real interpreter.
Since we're establishing commissions as the even-higher-level graph
concept of note, that means we're pretty sure we're not going to need
any of our skylark scripts to emit more than one basting at a time,
and thus I'm going to establish a convention of producing your result
by assigning it to a well-known global variable by the time your
script returns.  And that works out pretty easily.

Signed-off-by: Eric Myhre <hash@exultant.us>