commission: pop in a DAG sort.
This is a little rough around the edges: some places the catalog/module
name is more semantically relevant, and in others, the release/version
name is relevant; and we're not yet completely consistent on that.
For now, all the code is operating on catalog name granularity only,
and we're going to expect to go back and touch it up with many more
release version references later before calling this whole system done.

But, we've got a total ordering for the commission graph now.

Which means we're *awfully* close to being able to start evaluating
pretty wicked massive amounts of stuff.  Whee!

Signed-off-by: Eric Myhre <hash@exultant.us>