e2e: put "std" lib skylark funcs in hardcode map.
This is a placeholder to help me get over some writer's block.
Final library handling will definitely not be so hardcode-y.

The demo material is about to start involving a fair number of build
nodes, so even if we lack a well-spec'd solution to library handling,
I definitely don't want to be copy-pasting this very much.

Signed-off-by: Eric Myhre <hash@exultant.us>