hitch,commission: comments about design questions.
This is rough, and kind of a funny-looking problem:
we have two levels and kinds of recursion to deal with as we prepare the commission.

One is the high level actual commission graph.

The second is the local exploration we have to do while feeling out
each potential immediate dependency while exploring a single node.

It's frustrating to try to come up with a clean design for this because
we're stuck between the trifecta of these rocks and hard places:

- The type-2 exploration loads some data that'll be useful to the big
  picture in just a moment, but...
- It's too much data to pass around in IPC sensibly;
- and we can't load it in the big picture and pass it down, either.

After meditating on it for a while, I can't see any way around this
other than just taking a DealWithIt/gordianKnot approach.
In this case, that means designing our APIs in the go-timeless-api
repo, and making sure they're serializable; and making sure even
if/when heft makes a shortcut in memory, it's still using the
"serializable"-ready APIs to do so, which should keep us closer to the
straight and narrow for future progress.

Signed-off-by: Eric Myhre <hash@exultant.us>