commission: let's *only* store import edges.
Things like "the set of release names referenced by the complete graph
for a selected catalogName", and "the set of imports which are
non-candidate and thus a 'seed set' that must be fetched"... are
useful, but can be projected from this graph easily enough.
So I'm indifferent towards computing all those convenience indexes
as we go unless something in profiling later tells us we really must.
Minimizing the amount of stuff we carry around will make this less of
a PITA to serialize, and generally should make simpler code.

Though it may still be tempting to keep some helpful stats like
"how many formulas/steps did this basting propose?" as we go...

So the struct stays.

This made it impossible to create the scenario about cycles in one of
the tests.  So that's commented out now.
We need a real interpreter that can do interesting things ASAP.
And I think I was mistaken to start mocking something there; the real
deal should be easy enough to use (and if it's not, that's a first
class problem that should be fixed, not avoided).

Signed-off-by: Eric Myhre <hash@exultant.us>