interpret: add interface specific to commissioning
The layer which returns a basting should be a little separated from
the rest of the interpreter setup.  And it seems like we'll be
evolving towards two modes of interpretation already: modules for
library functions versus the heft scripts have different levels of
purity (libraries are restricted to being pure and nilipotent; heft
scripts still nilipotent, but allowed to read things).

I might be going a little interface-crazy here (we're already backing
away from thinking mocking at this level is sensible), but it seems
like the interface might be useful for adding a caching layer.
And maybe it'll actually even turn out correct for a non-heft concept
of planner script?  But again, trying to keep our hopes constrained
and feet on the ground, there.

Signed-off-by: Eric Myhre <hash@exultant.us>