all: use real interpreter.  several mocks removed!
Lots of things happening in this commit:

- destroyed several "mock" interpreters that were their own unique snowflakeish
placeholders; we have a real one now.

- and really, it *is* the real one.  Our commission fixtures now exercise the
skylark interpreter, and our core API types and functions in skylark.

- the "stdlib" functions moved again.  I suppose they're just going to keep
floating around for a while, until everything else settles.

- hitch view func injections are still todo.  They're running around with nils
ecause we have no skylark builtins for making that info accessible yet either.

Some of the commission tests get a tad verbose.  That can probably be fixed up
with some template strings for the heft scripts, and that should give us much
better exercise in general than the hacky mocked interpreter, and do it with
a minimum of fuss.

The e2e tests in the placeholder 'river' package are increasingly... left
without much interesting to do.  Good!  All that code was just meant to
be a forcing function and exploratory work; not meant to live on forever.

I think now we just need to propose hitch funcs for use in the interpreter,
and we'll have almost the bulk of this work done.  (There's lots of other
glue around e.g. URL handling, and fixing skyform basting to do exports, etc
that's lurking around.  But despite being death by papercuts, they're
nonblocking, so this progress still feels like we're getting "close".)

Signed-off-by: Eric Myhre <hash@exultant.us>