e2e: we *can* test graph sort now, you know...
Real bastings should most definitely *not* hardcode the "candidate" version,
because if you do that to yourself, what's the point?  You'll then *never*
generate a basting that's actually publicly releasable.

But we can do it, for testing and fooling around.  And it does trigger the
conditions for whether or not we demand that the builds recurse.

Finally, note that while it's hard to test this with this "river" dir full of
happy-path hijinx... if you put "candidate" in the busybase reference?
"cannot find planner script for making candidate releases of ..." -- correct!

Signed-off-by: Eric Myhre <hash@exultant.us>